Data type change

+++ GlobalGroupMembers/main.cpp
@@ -221,7 +221,7 @@ VOID EnumGlobalGroupMembers(std::vector <std::wstring> vwsArgs) {
 		return;
 	}
 
-	__int64 i = 0;
+	size_t i = 0;
 	while (i < vwsArgs.size()) {
 		if (vwsArgs[i] == L"-h") {
 			PrintUsage();
@@ -281,7 +281,7 @@ VOID EnumGlobalGroupMembers(std::vector <std::wstring> vwsArgs) {
 		return;
 	}
 
-	UINT j = 0;
+	size_t j = 0;
 	while (j < vTargets.size()) {
 		std::wcout << L"[" << j + 1 << L":" << vTargets.size() << L"] " << vTargets[j] << L"...";
